week-3/Day-2-Tuesday/lab-09-python-etl-pipeline-development/lab-9a-python-etl/requirements.txt
# Database setup dependencies
pyodbc>=4.0.39
python-dotenv>=1.0.0
